Finalise checkout 3ds service public interface
